+++ src/gui/dialogs/statistics_dialog.cpp
@@ -71,7 +71,11 @@ void statistics_dialog::pre_show(window& window)
 	// Set title
 	//
 	label& title = find_widget<label>(&window, "title", false);
-	title.set_label((formatter() << title.get_label() << " (" << current_team_.side_name() << ")").str());
+
+	// Show side name in parentheses if it is set, otherwise show nothing
+	std::string formatted_side_name = current_team_.side_name() == "" ? "" : "(" + current_team_.side_name() + ")";
+
+	title.set_label((formatter() << title.get_label() << formatted_side_name).str());
 
 	//
 	// Set up scenario menu